requirements.txt
streamlit
streamlit-option-menu
pandas
psycopg2-binary
folium
streamlit_folium
numpy
scipy
geopandas
shapely
scikit-learn
leafmap
ipython 
ipyleaflet 
ipywidgets  
branca   
rasterio
pyproj
google-cloud-aiplatform
google-generativeai
tabulate
geemap
earthengine-api